requirements.txt
langchain
ipykernel
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
faiss-cpu